Print the Session Information Markdown

Session information can be printed in a rmarkdown report. This function printss the necessary text to make it look nice in the markdown.

cat_session_markdown(f = stdout())

Arguments

f

File object to append markdown to (stdout by default).

Value

A string representing the rmarkdown needed for session information.

Details

Session information is an important part of data science reports, since it provides the package information and system configuration when the report was originally generated.

This function prints to the screen the necessary rmarkdown that can be included in the source of a report. When the report is rendered, the session information will be included.

Note

This function does not generate a report, only the markdown needed to generate such a report during render.

Examples

cat_session_markdown()
#> 
#> <!---
#> ATTENTION: The below text should remain at the bottom of your markdown
#> as it provides tables about the runtime environment. Leave
#> everything from the start of this paragraph down to the end of the
#> document.
#> --->
#> \newpage
#> # Session Information
#> The information in this section is included in the report to facilitate
#> reproducible research. The specific environment under which the code was
#> run is detailed.
#> ```{r session-information, echo=FALSE, results='asis'}
#> dsreportr::print_session_information()
#> ```
#>