List available templates within a package.

Templates are defined using a particular structure (for RStudio). This function lists the available templates from a named package.

list_templates(pkg)

Arguments

pkg

A package to enumerate templates within.

Value

A simple kable representing the available templates.

Details

Templates are defined by a couple of key files in particular locations within a package. One notable file is the template.yaml file which provides a name and description of the template.

This function looks in the package and reports on all templates that have a template.yaml file in the rmarkdown/templates directory. It reports on the template name, pretty name and description.

Note

This function is vectorized, so pkg can be a vector of packages.

Examples

if (FALSE) {
list_template("dsreportr")
## template   name                              description
## ---------  --------------------------------  ------------------------------------
## ds_docx    Data Science DOCX Report Output   A Generic Data Science DOCX report.
## ds_pdf     Data Science PDF Report Output    A Generic Data Science PDF report.

list_template(c("dsreportr","qsctemplates"))
## template   name                              description
## ---------  --------------------------------  -------------------------------------------------
## ds_docx    Data Science DOCX Report Output   A Generic Data Science DOCX report.
## ds_pdf     Data Science PDF Report Output    A Generic Data Science PDF report.
## qsc_pdf    QSC PDF Template                  A template for the U54 QSC based on dsreportr.
## u54_pdf    U54 PDF Report Template           A template for U54 reporting (not QSC specific).
}